requirements.txt
asgiref==3.10.0
boto3==1.37.37
botocore==1.37.38
Django==5.2
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
et_xmlfile==2.0.0
geographiclib==2.1
geopy==2.4.1
jmespath==1.0.1
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
